feat: DECSLPP/DECSNLS dynamic height + XTWINOPS case 8 + height on the wire (#39)
* feat: DECSLPP/DECSNLS dynamic height + XTWINOPS case 8 + height on the wire

* fix: resize-geometry bounds + xterm-correct grow/shrink anchoring and margin reset (review PR #39)

* fix: review nits — SGRTest stride, shift-trap comments on mask loops, XTWINOPS bare-seq test (PR #39 re-review)

* fix: re-review nits — fixture sync, clamp comment, MAX_HEIGHT structural guard (PR #39 round 3)

* fix: review round 4 — float mouse mapping, clamp wire scroll-window, blink-loop frame capture, 64-row mask test (PR #39)

* fix: review round 5 — grow view anchoring (lrd>=height invariant), alt-active grow cursor, overlay render row guards (PR #39)

* fix: review round 6 -- torn-read guard bound-checks the wrong arrays (PR #39)

The guard added for the resize/render race (§36 M4) checked only
activeStyles.length before indexing terminal.buffer/altBuffer (and,
via getForegroundColor/resolveColor, colors/colorsBackground). A frame
observing the new (longer) styles array with the still-old (shorter)
buffer/color arrays passed the bounds check and then read past them.

Bound-check against every array actually indexed in the row instead of
just styles.

---------

+++ src/main/java/li/cil/oc2/common/vm/terminal/render/RendererModel.java
@@ -1,9 +1,9 @@
 package li.cil.oc2.common.vm.terminal.render;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public interface RendererModel {
-    AtomicInteger getDirtyMask();
+    AtomicLong getDirtyMask();
 
     void close();
 }
